Move InitializationSettings to itws own header
